requirements.txt
aiogram>=3.0.0
aiohttp
python-dotenv>=1.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
alembic>=1.13.0
aiohttp-socks
client3x
pytest>=7.0.0
pytest-asyncio>=0.21.0
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.0
python-multipart>=0.0.6
passlib>=1.7.4
python-jose>=3.3.0
pydantic-settings>=2.0.0
pydantic>=2.0.0
yookassa>=3.3.0
cryptography>=41.0.0